requirements-voice.txt
# Core dependencies
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.104.0

# Date/Time parsing
python-dateutil>=2.8.2

# UI/Display
prettytable==3.10.0

# Voice Recognition
SpeechRecognition==3.10.1
# pyaudio==0.2.13  # Commented out due to Python 3.13 compatibility issues

# Voice Output (Text-to-Speech)
pyttsx3==2.90

# GUI Development
# tkinter comes with Python, not available via pip
tkcalendar>=1.6.1

# AI/ChatGPT Integration
openai>=1.0.0
python-dotenv>=1.0.0

# Web Dashboard
Flask>=2.3.0
Werkzeug>=2.3.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Natural language date parsing
dateparser>=1.1.4
parsedatetime>=2.6
